Kill torchtext for good
Remove the last remaining bits to .tasks (dataset download utilities)
and to .data (Vocab classes, word vectors)
