+++ decanlp/tasks/dataset.py
@@ -1,11 +1,10 @@
-import io
 import os
 import zipfile
 import tarfile
 
 import torch.utils.data
 
-from ..utils import download_from_url
+from .utils import download_from_url
 
 
 class Dataset(torch.utils.data.Dataset):
@@ -122,3 +121,15 @@ class Dataset(torch.utils.data.Dataset):
                         tar.extractall(path=path, members=dirs)
 
         return os.path.join(path, cls.dirname)
+
+
+def interleave_keys(a, b):
+    """Interleave bits from two sort keys to form a joint sort key.
+
+    Examples that are similar in both of the provided keys will have similar
+    values for the key defined by this function. Useful for tasks with two
+    text fields like machine translation or natural language inference.
+    """
+    def interleave(args):
+        return ''.join([x for t in zip(*args) for x in t])
+    return int(''.join(interleave(format(x, '016b') for x in (a, b))), base=2)
