--- decanlp/text/__init__.py
@@ -1,3 +1,0 @@
-__version__ = '0.2.1'
-
-__all__ = ['data', 'utils']
