--- decanlp/text/data/__init__.py
@@ -1,4 +1,0 @@
-from .dataset import Dataset
-from .utils import interleave_keys
-
-__all__ = ["Dataset", "interleave_keys"]
